[OPT] change package name to io.john6.base, move JWheelPickerInfo to .bean
