+++ JWheelPicker/proguard-rules.pro
@@ -22,7 +22,7 @@
 -keepclassmembers public class * extends androidx.lifecycle.ViewModel {
    public <init>(...);
 }
--keepclassmembers public class * extends io.john6.johnbase.compose.picker.dialog.multiple.IMultipleJPickerAdapter {
+-keepclassmembers public class * extends io.john6.base.compose.picker.dialog.multiple.IMultipleJPickerAdapter {
    public <init>(...);
 #   public <init>(android.os.Bundle);
 #   public <init>();
