+++ JWheelPicker/src/main/kotlin/io/john6/base/compose/picker/dialog/multiple/JMultiplePickerViewModel.kt
@@ -1,13 +1,16 @@
-package io.john6.johnbase.compose.picker.dialog.multiple
+package io.john6.base.compose.picker.dialog.multiple
 
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
-import io.john6.johnbase.compose.picker.bean.JWheelPickerItemInfo
+import io.john6.base.compose.picker.bean.JWheelPickerItemInfo
 
 class JMultiplePickerViewModel(savedStateHandle: SavedStateHandle) :ViewModel(){
     val requiredData: JMultiPickerDialogData = savedStateHandle.get<JMultiPickerDialogData>("data")!!
 
-    val mMultipleJPickerAdapter = IMultipleJPickerAdapter.create(requiredData.adapterClass, requiredData.adapterParamsAsBundle)
+    val mMultipleJPickerAdapter = IMultipleJPickerAdapter.create(
+        requiredData.adapterClass,
+        requiredData.adapterParamsAsBundle
+    )
 
     /**
      * 当前选中的 item 信息，初始为 [JWheelPickerItemInfo.EMPTY]
