+++ build.gradle.kts
@@ -1,6 +1,3 @@
-buildscript {
-    val agp_version by extra("8.0.2")
-}
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
     id("com.android.application") version "8.1.1" apply false
